Refactor analysis logic into focused modules with macOS separation
This change restructures the analysis layer into smaller, focused files while preserving existing behavior and UI contracts. The main AnalysisLogic type remains the facade used by the view, with file selection, analysis flow, USB handling, lifecycle, and notifications split into dedicated units. macOS-specific analysis responsibilities were moved into a dedicated macOS logic folder and renamed consistently to prepare cleanly for future Linux-specific counterparts without altering runtime rules.
