+++ .github/workflows/development_build_debug.yml
@@ -1,0 +1,33 @@
+name: Debug Build (development)
+
+on:
+  push:
+    branches:
+      - development
+  pull_request:
+    branches:
+      - development
+
+concurrency:
+  group: debug-build-${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
+jobs:
+  debug-build:
+    name: macUSB - Debug Build
+    runs-on: macos-latest
+    timeout-minutes: 30
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Build macUSB (Debug)
+        run: |
+          xcodebuild \
+            -project macUSB.xcodeproj \
+            -scheme macUSB \
+            -configuration Debug \
+            -destination 'platform=macOS' \
+            CODE_SIGNING_ALLOWED=NO \
+            build
