+++ macUSB/Features/Analysis/AnalysisNotifications.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+extension Notification.Name {
+    static let macUSBResetToStart = Notification.Name("macUSB.resetToStart")
+    static let macUSBStartTigerMultiDVD = Notification.Name("macUSB.startTigerMultiDVD")
+    static let macUSBDebugGoToBigSurSummary = Notification.Name("macUSB.debugGoToBigSurSummary")
+    static let macUSBDebugGoToTigerSummary = Notification.Name("macUSB.debugGoToTigerSummary")
+    static let macUSBNavigateToAnalysis = Notification.Name("macUSB.navigateToAnalysis")
+    static let macUSBApplyPendingDownloaderInstaller = Notification.Name("macUSB.applyPendingDownloaderInstaller")
+}
